Detectron2/requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
anyio==3.3.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
deprecation==2.1.0
-e git+https://github.com/facebookresearch/detectron2@b669a88fd174decf9b5a2b3e34e0d8305eaabd4b#egg=detectron2
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210804
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
hydra-core==1.1.0
idna==3.2
importlib-resources==5.2.2
iopath==0.1.9
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.10.4
jupyter-server==1.10.2
jupyterlab==3.1.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271903025/work
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.0
open3d==0.13.0
opencv-python==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1625663293593/work
portalocker==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrealsense2==2.48.0.3381
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==225
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
Send2Trash==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tomlkit==0.7.2
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
yacs==0.1.8
zipp==3.5.0
